Keep accurate track of window changes in workspace buttons for PagerShowPreview.  Only redraw affected workspace buttons, which reduces the total number of workspace button redraws, especially for larges number of workspace buttons.

+++ src/wmwinlist.cc
@@ -267,7 +267,7 @@ void WindowListBox::enableCommands(YMenu *popup) {
             traytoggle |= notbit(frame->frameOptions(), YFrameWindow::foIgnoreTaskBar);
             closable |= (frame->canClose());
 
-            long ws = frame->getWorkspace();
+            int ws = frame->getWorkspace();
             if (workspace == AllWorkspaces) {
                 workspace = ws;
                 sameWorkspace = true;
